% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_map_state.R
\name{fars_map_state}
\alias{fars_map_state}
\title{Generate a map of a state showing locations of accidents throughout a given year.}
\usage{
fars_map_state(state.num, year)
}
\arguments{
\item{state.num}{a string or numeric representing the coded value for
a state.}

\item{year}{a string or numeric representing the year for which we want to load accident data.}
}
\value{
a map of a state showing locations of the accidents for a given year
}
\description{
This function takes in a numbered code for a state and a year as arguments,
generates a file name, and reads in data from the file into a data frame.
}
\details{
The function checks to make sure the state number is valid (returns a message "invalid state number" otherwise),
and creates a new dataframe consisting only of accidents from that state.

This function will also result in an error if an invalid value for year is given.

If there are no accidents in the state for that year, it returns a message to the user.

The function then checks for longitude values greater than 900 and
latitude values greater than 90 and turns them into NAs.

The function then generates the map using longitude and latitude values for the points.
}
\examples{
fars_map_state(05, 2005)

state <- 05
year <- 2005
fars_map_state(state, year)

}
