% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_read.R
\name{fars_read}
\alias{fars_read}
\title{Read a FARS file}
\usage{
fars_read(filename)
}
\arguments{
\item{filename}{a character string representing the name of the file to be read}
}
\value{
This function returns a data frame consisting of
 the data read in from the given file name, or a message that
 the file does not exist
}
\description{
This functions checks to see if a given file exists,
and if it does, it reads it in and returns the data
as a data frame. If the file does not exist, it returns
a message that the file does not exist
}
\examples{
fars_read("data.csv")

name <- "data.csv"
fars_read(name)

}
