% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_read_years.R
\name{fars_read_years}
\alias{fars_read_years}
\title{Read in files for multiple years}
\usage{
fars_read_years(years)
}
\arguments{
\item{years}{a vector of strings or numerics, each in "yyyy" format.}
}
\value{
A list of dataframes, one for each year in years
}
\description{
This function takes a vector of strings or numerics,
uses lapply to call make_filename() for each year in the vector,
and attempts to read in data for each file name with fars_read.
}
\details{
The function then uses mutate() to add a column specifying the year
and returns a dataframe consisting of the month and year of each row.

If fars_read() fails to read in data for a particular year,
it prints a warning to the user and returns NULL.
}
\examples{
fars_read_years(c(2016, 2017, 2018))

years <- 2005:2014
data <- fars_read_years(years)

}
