% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_summarize_years_doc.R
\name{fars_summarize_years}
\alias{fars_summarize_years}
\title{Summarize the number of accidents in each month}
\usage{
fars_summarize_years(years)
}
\arguments{
\item{years}{a vector of strings or numerics, each in "yyyy" format.}
}
\value{
a dataframe consisting of the number of accidents of each month,
with the summarized years as the column names.
}
\description{
This function takes a vector of strings or numerics representing
which years will be summarized. The vector is passed to
fars_read_years to create a list of dataframes.
}
\details{
The function then calls bind_rows to combine the dataframes into one.
The single dataframe is then passed into group_by and summarize to
obtain the total number of accidents each month.

Finally, the summarized dataframe is passed to spread to create a dataframe
where the columns are each year, the rows are the months, and the values are the
number of accidents each month for each year.

If years does not consist of strings or numerics representing the years to summarize,
an error may result.
}
\examples{
fars_summarize_years(c(2005, 2010, 2015))

years <- 2000:2009
df <- fars_summarize_years(years)

}
