% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_filename.R
\name{make_filename}
\alias{make_filename}
\title{Create the proper file name}
\usage{
make_filename(year)
}
\arguments{
\item{year}{A string or numeric in "yyyy" format}
}
\value{
This function returns a string generated using the year provided
}
\description{
This function takes a year and uses it to create a string representing a file name.
Will result in an error if the string cannot be coerced to an integer
}
\examples{
## make_filename("2015")

## year <- 2015
## fileName <- make_filename(year)

}
